refactor: move make_api_request.py into call_brain_api.py
