+++ backend/modules/brain/service/call_brain_api.py
@@ -1,14 +1,57 @@
 from uuid import UUID
+import requests
+from logger import get_logger
+logger = get_logger(__name__)
 
 from fastapi import HTTPException
 
 from modules.brain.entity.api_brain_definition_entity import ApiBrainDefinitionSchema
-from utils.make_api_request import get_api_call_response_as_text
 from modules.brain.service.api_brain_definition_service import ApiBrainDefinitionService
 from modules.brain.service.brain_service import BrainService
 
 brain_service = BrainService()
 api_brain_definition_service = ApiBrainDefinitionService()
+
+
+def get_api_call_response_as_text(
+    method, api_url, params, search_params, secrets
+) -> str:
+    headers = {}
+
+    api_url_with_search_params = api_url
+    if search_params:
+        api_url_with_search_params += "?"
+        for search_param in search_params:
+            api_url_with_search_params += (
+                f"{search_param}={search_params[search_param]}&"
+            )
+
+    for secret in secrets:
+        headers[secret] = secrets[secret]
+
+    try:
+        if method in ["GET", "DELETE"]:
+            response = requests.request(
+                method,
+                url=api_url_with_search_params,
+                params=params or None,
+                headers=headers or None,
+            )
+        elif method in ["POST", "PUT", "PATCH"]:
+            response = requests.request(
+                method,
+                url=api_url_with_search_params,
+                json=params or None,
+                headers=headers or None,
+            )
+        else:
+            raise ValueError(f"Invalid method: {method}")
+
+        return response.text
+
+    except Exception as e:
+        logger.error(f"Error calling API: {e}")
+        return None
 
 
 def extract_api_brain_definition_values_from_llm_output(
